+++ src/LIN3S/AdminBundle/DependencyInjection/Configuration.php
@@ -19,7 +19,7 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
  *
  * @author Gorka Laucirica <gorka.lauzirka@gmail.com>
  */
-class Configuration implements ConfigurationInterface
+final class Configuration implements ConfigurationInterface
 {
     /**
      * {@inheritdoc}